+++ engine/src/lib.rs
@@ -1,7 +1,7 @@
+pub mod cache;
 pub mod catalog;
 pub mod files_manager;
 pub mod paged_file;
-pub mod poc_cache;
 
 pub fn add(left: u64, right: u64) -> u64 {
     left + right
